Adicionando o swagger e mudando os ids para long

+++ src/main/java/com/api/solset/controller/BudgetController.java
@@ -30,7 +30,7 @@ public class BudgetController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable Integer id){
+    public ResponseEntity<?> delete(@PathVariable Long id){
         budgetService.delete(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
