+++ src/main/java/com/api/solset/repository/BudgetRepository.java
@@ -4,6 +4,6 @@ import com.api.solset.model.Budget;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface BudgetRepository extends JpaRepository<Budget, Integer> {
-    List<Budget> findByInstallationId(Integer installationId);
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
+    List<Budget> findByInstallationId(Long installationId);
 }
