+++ src/main/java/com/api/solset/repository/ClientRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ClientRepository extends JpaRepository<Client, Integer> {
-        List<Client> findByUserId(Integer id);
+public interface ClientRepository extends JpaRepository<Client, Long> {
+        List<Client> findByUserId(Long id);
 }
