+++ src/main/java/com/api/solset/repository/InstallationRepository.java
@@ -4,6 +4,6 @@ import com.api.solset.model.Installation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface InstallationRepository extends JpaRepository<Installation, Integer> {
-    List<Installation> findByClientId(Integer clientId);
+public interface InstallationRepository extends JpaRepository<Installation, Long> {
+    List<Installation> findByClientId(Long clientId);
 }
