+++ src/main/java/com/api/solset/service/UserService.java
@@ -37,7 +37,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void delete(Integer id){
+    public void delete(Long id){
         userRepository.deleteById(id);
     }
 }
